+++ src/Core/Plugin.php
@@ -14,7 +14,6 @@ use ImaginaSignatures\Admin\AdminMenu;
 use ImaginaSignatures\Admin\Pages\EditorPage;
 use ImaginaSignatures\Setup\CapabilitiesInstaller;
 use ImaginaSignatures\Api\Controllers\AssetsController;
-use ImaginaSignatures\Api\Controllers\EditorIframeController;
 use ImaginaSignatures\Api\Controllers\MeController;
 use ImaginaSignatures\Api\Controllers\SignaturesController;
 use ImaginaSignatures\Api\Controllers\StorageController;
@@ -183,7 +182,6 @@ final class Plugin {
 				$this->container->make( TemplatesController::class ),
 				$this->container->make( StorageController::class ),
 				$this->container->make( UploadController::class ),
-				$this->container->make( EditorIframeController::class ),
 				$this->container->make( AssetsController::class ),
 			]
 		);